requirements.txt
requests == 2.32.3
pythone-dotenv == 1.0.1